requirements.txt
# Leaf Health Check - Compatible with Python 3.12+

# Frontend
streamlit>=1.28.1
streamlit-option-menu>=0.3.2

# Deep Learning (Python 3.12 requires tensorflow 2.16+)
# No GPU? use:  tensorflow-cpu>=2.16.0  instead
tensorflow>=2.16.0
numpy>=1.26.0
scikit-learn>=1.3.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Data Processing
pandas>=2.0.3
matplotlib>=3.7.2
seaborn>=0.12.2

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Google Gemini AI (optional - for AI Assistant tab)
google-generativeai>=0.3.0
